fix(runtime): address cloudflare adapter review
